Refactor MainActivity and update dependencies
This commit refactors `MainActivity.kt` by moving the drawer initialization logic into the `openMain()` function. This improves the organization and readability of the code.

Additionally, the `navigation-runtime-ktx` dependency has been removed as it is already included as a transitive dependency of `navigation-fragment-ktx`.

The ID of the `DrawerNavigationView` in `activity_main.xml` has also been changed from `navigation_view` to `navigationView` for consistency.